+++ examples/multi/src/app.rs
@@ -2,34 +2,44 @@ use zng::prelude::*;
 
 #[allow(unused)]
 pub fn run() {
-    APP.defaults().run_window(async {
-        let count = var(0u32);
+    let app = APP.defaults();
+    CONFIG.load(zng::config::JsonConfig::sync(zng::env::config("config.json")));
+
+    let res = zng::env::res("my-res.txt");
+    match std::fs::read_to_string(&res) {
+        Ok(res) => tracing::info!("resource read ok: {res}"),
+        Err(e) => tracing::error!("resource read error, cannot read '{}', {e}", res.display()),
+    }
+
+    app.run_window(async {
+        let count = CONFIG.get("count", 0u32);
         Window! {
             child = Button! {
                 style_fn = zng::button::LightStyle!();
                 child_align = Align::TOP;
                 child = Text! {
-                    txt = count.map(|&c| {
-                        let mut txt = formatx!("HELLO {}!", std::env::consts::OS.to_uppercase());
-                        if c > 0 {
-                            use std::fmt::Write;
-                            write!(&mut txt, "\n{c}");
-                        }
-                        txt
-                    });
+                    txt = count.map(|&c| formatx!("HELLO {}!\n\n{c}", std::env::consts::OS.to_uppercase()));
                     txt_align = Align::CENTER;
-                    layout::padding = 20;
+                    layout::padding = 40;
                     font_family = "fantasy";
                     font_size = 2.em();
                     font_weight = FontWeight::BOLD;
                     font_style = FontStyle::Italic;
 
                 };
-                on_click = hn!(|_| {
-                    let c = count.get() + 1;
+                on_click = hn!(count, |_| {
+                    let c = count.get().wrapping_add(1);
                     tracing::info!("Clicked {c} times!");
                     count.set(c);
                 });
+                context_menu = ContextMenu!(ui_vec![
+                    Button! {
+                        child = Text!("Reset");
+                        on_click = hn!(|_| {
+                            count.set(0u32);
+                        });
+                    }
+                ]);
             };
             color_scheme = color::ColorScheme::Dark;
             widget::background_gradient = {